+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - generate types for utils subdirectory (#287)
 - export gobject utils in Utils
+- bind service methods
 
 ## Fixes